Add files via upload
In C++, a friend function is a function that is not a member of a class but is granted access to its private and protected members. This means that a friend function can access private and protected members of a class, even though it is not a member of that class.